# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NER-LogiSense Intelligence Platform",
    "description": "\n# NER-LogiSense Backend API\n### AI-powered Smart Logistics and Accessibility Intelligence Platform for India's North Eastern Region (NER)\n\n**Core Capabilities**:\n- Real-time road, bridge, and transport accessibility monitoring\n- Landslide and flood disruption prediction using environmental feature models\n- AI-based multi-criteria route optimization ($Cost = \\alpha \\cdot T + \\beta \\cdot R + \\gamma \\cdot C$)\n- High-throughput GPS tracking of vehicles carrying medical oxygen, cryo-vaccines, foodgrains, and relief supplies\n- Offline-first field hazard reporting with idempotency and conflict resolution\n- Multilingual automated alerts across 9 regional languages\n- WebSocket telemetry streaming\n    ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/alerts": {
      "post": {
        "tags": [
          "Alerts & Tactical Emergency Coordination"
        ],
        "summary": "Create Alert",
        "description": "Create automated or manual alert with AI deduplication check.",
        "operationId": "create_alert_api_v1_alerts_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Alerts & Tactical Emergency Coordination"
        ],
        "summary": "List Alerts",
        "description": "List operational alerts.",
        "operationId": "list_alerts_api_v1_alerts_get",
        "parameters": [
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "ACTIVE",
              "title": "Status Filter"
            }
          },
          {
            "name": "severity_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlertResponse"
                  },
                  "title": "Response List Alerts Api V1 Alerts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alerts/{id}": {
      "patch": {
        "tags": [
          "Alerts & Tactical Emergency Coordination"
        ],
        "summary": "Update Alert",
        "description": "Acknowledge or resolve an active alert.",
        "operationId": "update_alert_api_v1_alerts__id__patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alerts/{id}/notify": {
      "post": {
        "tags": [
          "Alerts & Tactical Emergency Coordination"
        ],
        "summary": "Dispatch Alert Notifications",
        "description": "Broadcast alert across regional channels (WebSocket, SMS, Push, WhatsApp).",
        "operationId": "dispatch_alert_notifications_api_v1_alerts__id__notify_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertNotifyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Dispatch Alert Notifications Api V1 Alerts  Id  Notify Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/notifications/{id}/read": {
      "patch": {
        "tags": [
          "Alerts & Tactical Emergency Coordination"
        ],
        "summary": "Mark Notification Read",
        "description": "Acknowledge notification read receipt.",
        "operationId": "mark_notification_read_api_v1_notifications__id__read_patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Mark Notification Read Api V1 Notifications  Id  Read Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notification-preferences": {
      "post": {
        "tags": [
          "Alerts & Tactical Emergency Coordination"
        ],
        "summary": "Save Notification Preferences",
        "description": "Save multilingual and channel preferences.",
        "operationId": "save_notification_preferences_api_v1_notification_preferences_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationPreferencesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Save Notification Preferences Api V1 Notification Preferences Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/emergency-events": {
      "post": {
        "tags": [
          "Alerts & Tactical Emergency Coordination"
        ],
        "summary": "Activate Emergency Event",
        "description": "Declare state-wide or district-wide emergency disaster mode (e.g. Monsoon Surge Level 3).",
        "operationId": "activate_emergency_event_api_v1_emergency_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmergencyEventCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyEventResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/emergency-events/{id}": {
      "patch": {
        "tags": [
          "Alerts & Tactical Emergency Coordination"
        ],
        "summary": "Update Emergency Event",
        "description": "Deactivate or update emergency disaster state.",
        "operationId": "update_emergency_event_api_v1_emergency_events__id__patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmergencyEventUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyEventResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlertCreate": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "severity": {
            "type": "string",
            "title": "Severity",
            "default": "HIGH"
          },
          "district_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District Id"
          },
          "road_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Road Id"
          },
          "location_desc": {
            "type": "string",
            "title": "Location Desc"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "type",
          "location_desc",
          "message"
        ],
        "title": "AlertCreate"
      },
      "AlertNotifyRequest": {
        "properties": {
          "channels": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channels",
            "default": [
              "WEBSOCKET",
              "PUSH",
              "SMS"
            ]
          },
          "target_districts": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Districts"
          },
          "priority": {
            "type": "string",
            "title": "Priority",
            "default": "HIGH"
          }
        },
        "type": "object",
        "title": "AlertNotifyRequest"
      },
      "AlertResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "district_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District Id"
          },
          "road_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Road Id"
          },
          "location_desc": {
            "type": "string",
            "title": "Location Desc"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "severity",
          "location_desc",
          "message",
          "status",
          "created_at"
        ],
        "title": "AlertResponse"
      },
...
      "EmergencyEventCreate": {
        "properties": {
          "event_code": {
            "type": "string",
            "title": "Event Code"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "default": "MONSOON_SURGE"
          },
          "level": {
            "type": "string",
            "title": "Level",
            "default": "LEVEL_3"
          },
          "affected_districts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Affected Districts"
          },
          "priority_corridors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Priority Corridors"
          }
        },
        "type": "object",
        "required": [
          "event_code",
          "title",
          "affected_districts",
          "priority_corridors"
        ],
        "title": "EmergencyEventCreate"
      },
      "EmergencyEventResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "event_code": {
            "type": "string",
            "title": "Event Code"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "level": {
            "type": "string",
            "title": "Level"
          },
          "affected_districts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Affected Districts"
          },
          "priority_corridors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Priority Corridors"
          },
          "activated_by": {
            "type": "string",
            "title": "Activated By"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "event_code",
          "title",
          "type",
          "level",
          "affected_districts",
          "priority_corridors",
          "activated_by",
          "start_time",
          "is_active"
        ],
        "title": "EmergencyEventResponse"
      },
      "EmergencyEventUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "affected_districts": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Affected Districts"
          },
          "priority_corridors": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority Corridors"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "EmergencyEventUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NotificationPreferencesRequest": {
        "properties": {
          "channels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Channels"
          },
          "districts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Districts"
          },
          "alert_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Alert Types"
          },
          "language": {
            "type": "string",
            "title": "Language"
          }
        },
        "type": "object",
        "required": [
          "channels",
          "districts",
          "alert_types",
          "language"
        ],
        "title": "NotificationPreferencesRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}